+++ package.json
@@ -69,8 +69,8 @@
     "clean": "rm -rf out/*",
     "generate:config": "tsx ./scripts/generate-config.mts",
     "build": "concurrently \"npm:build:extension\" \"npm:build:tests\" \"npm:typecheck\" --prefixColors=auto",
-    "build:extension": "node esbuild.config.mjs",
-    "build:tests": "node esbuild.config.mjs --tests",
+    "build:extension": "tsx esbuild.config.mts",
+    "build:tests": "npm run build:extension -- --tests",
     "typecheck": "tsc --noEmit",
     "typecheck:prod": "npm run typecheck -- -p tsconfig.prod.json",
     "watch": "concurrently \"npm:watch:extension\" \"npm:watch:tests\" \"npm:watch:typecheck\" --prefixColors=auto",
@@ -79,7 +79,7 @@
     "watch:tests": "npm run build:tests -- --watch",
     "watch:typecheck": "npm run typecheck -- --watch --preserveWatchOutput",
     "watch:typecheck:prod": "npm run typecheck:prod -- --watch --preserveWatchOutput",
-    "lint": "eslint src/ scripts/",
+    "lint": "npx --node-options='--experimental-strip-types' eslint src/ scripts/ *.mts --flag unstable_native_nodejs_ts_config",
     "lint:fix": "npm run lint -- --fix",
     "format": "prettier . --write",
     "format:check": "prettier . --check",
@@ -95,10 +95,11 @@
   },
   "devDependencies": {
     "@cspell/eslint-plugin": "^8.17.3",
-    "@eslint/js": "^9.15.0",
+    "@eslint/js": "^9.33.0",
     "@stylistic/eslint-plugin": "^5.2.2",
     "@types/chai": "^4.3.20",
     "@types/chai-as-promised": "^7.1.8",
+    "@types/eslint-config-prettier": "^6.11.3",
     "@types/mocha": "^10.0.10",
     "@types/node": "22.x",
     "@types/node-fetch": "^2.6.12",
@@ -116,8 +117,8 @@
     "dotenv": "^16.5.0",
     "esbuild": "^0.25.2",
     "esbuild-node-externals": "^1.18.0",
-    "eslint": "^9.15.0",
-    "eslint-config-prettier": "^9.1.0",
+    "eslint": "^9.33.0",
+    "eslint-config-prettier": "^9.1.2",
     "eslint-plugin-check-file": "^3.1.0",
     "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-license-header": "^0.8.0",
