+++ src/test/e2e.mocharc.js
@@ -1,3 +1,9 @@
+/**
+ * @license
+ * Copyright 2025 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 const isDebugMode = process.argv.includes("--debug");
 module.exports = {
   timeout: isDebugMode ? 99999999 : "2m",
